fix(compose): resolve alignment, model type casts, and foundation opt-in annotations

+++ app/src/main/java/com/ndikanime/app/presentation/screens/chat/GlobalChatScreen.kt
@@ -63,7 +63,7 @@ fun GlobalChatScreen(
             id = System.currentTimeMillis().toString(),
             senderName = userProfile?.name ?: "Kamu",
             senderAvatar = userProfile?.picture,
-            senderLevel = userProfile?.level ?: 1,
+            senderLevel = (userProfile?.level ?: 1L).toInt(),
             text = inputText.trim()
         )
         messages.add(newMsg)
